Feat: 데이터 플랫폼 전송 카프카로 변경
